vorax_repo/pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.2.0",
    "flask>=3.1.0",
    "flask-sqlalchemy>=3.1.1",
    "gitpython>=3.1.44",
    "gunicorn>=23.0.0",
    "kaggle>=1.7.4.2",
    "numpy>=2.2.4",
    "psutil>=7.0.0",
    "psycopg2-binary>=2.9.10",
]
